refactor: Organize external and internal extension into its repositories
