+++ packages/aws-lambda-otel-extension/opt/extensions/otel-extension
@@ -8,4 +8,4 @@ OWN_FILENAME="$(basename $0)"
 LAMBDA_EXTENSION_NAME="$OWN_FILENAME" # (external) extension name has to match the filename
 
 echo "${LAMBDA_EXTENSION_NAME}  launching extension"
-exec "/opt/${LAMBDA_EXTENSION_NAME}/index.js"
+exec "/opt/${LAMBDA_EXTENSION_NAME}/external/index.js"
