+++ packages/aws-lambda-otel-extension/opt/otel-extension/external/lambda-apis/extensions-api.js
