+++ packages/aws-lambda-otel-extension/opt/otel-extension/external/lambda-apis/logs-api.js
