+++ packages/aws-lambda-otel-extension/opt/otel-extension/external/lambda-apis/runtime-api.js
