+++ packages/aws-lambda-otel-extension/opt/otel-extension/external/proto/common.proto
