+++ packages/aws-lambda-otel-extension/opt/otel-extension/external/proto/metrics.proto
