+++ packages/aws-lambda-otel-extension/opt/otel-extension/internal/event-detection/event-types/cloud-watch-log.js
