+++ packages/aws-lambda-otel-extension/opt/otel-extension/internal/event-detection/event-types/firehose.js
