+++ packages/aws-lambda-otel-extension/opt/otel-extension/internal/event-detection/event-types/sns.js
