+++ packages/aws-lambda-otel-extension/opt/otel-extension/internal/index.js
@@ -24,9 +24,9 @@ const { PgInstrumentation } = require('@opentelemetry/instrumentation-pg');
 const { RedisInstrumentation } = require('@opentelemetry/instrumentation-redis');
 const { diag, DiagConsoleLogger } = require('@opentelemetry/api');
 const fetch = require('node-fetch');
-const { logMessage, OTEL_SERVER_PORT, OTEL_SERVER_HOST } = require('./helper');
-const SlsSpanProcessor = require('./span.processor');
-const { detectEventType } = require('./eventDetection');
+const { logMessage, OTEL_SERVER_PORT, OTEL_SERVER_HOST } = require('../lib/helper');
+const SlsSpanProcessor = require('./span-processor');
+const { detectEventType } = require('./event-detection');
 
 const logLevel = getEnv().OTEL_LOG_LEVEL;
 diag.setLogger(new DiagConsoleLogger(), logLevel);
