+++ packages/aws-lambda-otel-extension/opt/otel-extension/internal/span-processor.js
