+++ packages/aws-lambda-otel-extension/opt/otel-extension/lib/helper.js
