+++ packages/aws-lambda-otel-extension/opt/otel-extension/otel-handler
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-export NODE_OPTIONS="${NODE_OPTIONS} --require /opt/otel-extension/wrapper.js"
+export NODE_OPTIONS="${NODE_OPTIONS} --require /opt/otel-extension/internal"
 
 if [[ $OTEL_RESOURCE_ATTRIBUTES != *"service.name="* ]]; then
   export OTEL_RESOURCE_ATTRIBUTES="service.name=${AWS_LAMBDA_FUNCTION_NAME},${OTEL_RESOURCE_ATTRIBUTES}"
